requirements.txt
azure-core==1.29.5
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-storage-file-share==12.15.0
azure-storage-queue==12.9.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Flask==3.0.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
